+++ src/components/Navbars/DemoNavbar.jsx
@@ -35,7 +35,7 @@ import {
   Input
 } from "reactstrap";
 
-import routes from "routes.js";
+import routes from "./../../routes/admin.js";
 
 class Header extends React.Component {
   state = {