+++ src/index.js
@@ -25,6 +25,7 @@ import "assets/scss/now-ui-dashboard.scss?v1.2.0";
 import "assets/css/demo.css";
 
 import AdminLayout from "layouts/Admin.jsx";
+import GeneralLayout from "layouts/General";
 
 const hist = createBrowserHistory();
 
@@ -32,6 +33,7 @@ ReactDOM.render(
   <Router history={hist}>
     <Switch>
       <Route path="/admin" render={props => <AdminLayout {...props} />} />
+      <Route path="/" render={props => <GeneralLayout {...props} />} />
       <Redirect to="/admin/dashboard" />
     </Switch>
   </Router>,