+++ src/routes/admin.js
@@ -15,14 +15,14 @@
 * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
 
 */
-import Dashboard from "views/Dashboard.jsx";
-import Notifications from "views/Notifications.jsx";
-import Icons from "views/Icons.jsx";
-import Typography from "views/Typography.jsx";
-import TableList from "views/TableList.jsx";
-import Maps from "views/Maps.jsx";
-import Upgrade from "views/Upgrade.jsx";
-import UserPage from "views/UserPage.jsx";
+import Dashboard from "./../views/Dashboard.jsx";
+import Notifications from "./../views/Notifications.jsx";
+import Icons from "./../views/Icons.jsx";
+import Typography from "./../views/Typography.jsx";
+import TableList from "./../views/TableList.jsx";
+import Maps from "./../views/Maps.jsx";
+import Upgrade from "./../views/Upgrade.jsx";
+import UserPage from "./../views/UserPage.jsx";
 
 var dashRoutes = [
   {